HDDS-12323. Improve container import failure cleanup flow

+++ hadoop-hdds/container-service/src/main/java/org/apache/hadoop/ozone/container/metadata/DatanodeStoreSchemaThreeImpl.java
@@ -17,10 +17,12 @@
  */
 package org.apache.hadoop.ozone.container.metadata;
 
+import com.google.common.annotations.VisibleForTesting;
 import org.apache.hadoop.hdds.StringUtils;
 import org.apache.hadoop.hdds.conf.ConfigurationSource;
 import org.apache.hadoop.hdds.protocol.proto.StorageContainerDatanodeProtocolProtos.DeletedBlocksTransaction;
 import org.apache.hadoop.hdds.upgrade.HDDSLayoutFeature;
+import org.apache.hadoop.hdds.utils.FaultInjector;
 import org.apache.hadoop.hdds.utils.MetadataKeyFilters;
 import org.apache.hadoop.hdds.utils.db.BatchOperation;
 import org.apache.hadoop.hdds.utils.db.FixedLengthStringCodec;
@@ -62,6 +64,7 @@ public class DatanodeStoreSchemaThreeImpl extends DatanodeStoreWithIncrementalCh
   public static final String DUMP_DIR = "db";
 
   private final Table<String, DeletedBlocksTransaction> deleteTransactionTable;
+  private static FaultInjector injector;
 
   public DatanodeStoreSchemaThreeImpl(ConfigurationSource config,
       String dbPath, boolean openReadOnly) throws IOException {
@@ -104,6 +107,10 @@ public class DatanodeStoreSchemaThreeImpl extends DatanodeStoreWithIncrementalCh
   }
 
   public void removeKVContainerData(long containerID) throws IOException {
+    if (getInjector() != null && getInjector().getException() != null
+        && getInjector().getTag().equalsIgnoreCase("removeKVContainerData")) {
+      throw (IOException) getInjector().getException();
+    }
     String prefix = getContainerKeyPrefix(containerID);
     try (BatchOperation batch = getBatchHandler().initBatchOperation()) {
       getMetadataTable().deleteBatchWithPrefix(batch, prefix);
@@ -134,6 +141,10 @@ public class DatanodeStoreSchemaThreeImpl extends DatanodeStoreWithIncrementalCh
 
   public void loadKVContainerData(File dumpDir)
       throws IOException {
+    if (getInjector() != null && getInjector().getException() != null
+        && getInjector().getTag().equalsIgnoreCase("loadKVContainerData")) {
+      throw (IOException) getInjector().getException();
+    }
     getMetadataTable().loadFromFile(
         getTableDumpFile(getMetadataTable(), dumpDir));
     getBlockDataTable().loadFromFile(
@@ -230,5 +241,15 @@ public class DatanodeStoreSchemaThreeImpl extends DatanodeStoreWithIncrementalCh
         }
       }
     }
+  }
+
+  @VisibleForTesting
+  public static FaultInjector getInjector() {
+    return injector;
+  }
+
+  @VisibleForTesting
+  public static void setInjector(FaultInjector instance) {
+    injector = instance;
   }
 }
